pyproject.toml
[project]
name = "reflective-resonance"
version = "0.1.0"
description = "Reflective Resonance art installation backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rawagents",
    "elevenlabs>=1.0.0",
]

[tool.uv.sources]
rawagents = { path = "/Users/tawab/Projects/MediaLab/rawagents", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
